refactor: share templates and crypto logic
