+++ restore-app/tsconfig.app.json
@@ -26,5 +26,5 @@
     "noUncheckedSideEffectImports": true
   },
   "types": ["node"],
-    "include": ["src", "src/custom.d.ts"],
+  "include": ["src", "src/custom.d.ts", "../shared"],
 }
